python/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
pywebview>=5.4
pydantic>=2.10
mss>=9.0
Pillow>=11.0
